move cmpreview to avoid issue when importing something from component use picture for gallery, review css to support responsive, reorganize gallery TODO fix hydration fliker effect
