--- src/commons/Gallery.client.module.css
@@ -1,127 +1,0 @@
-.galleryDialog {
-	--luxe-dialog-opacity: 0.9;
-}
-
-.galleryItemMain {
-  margin-bottom: 0;
-  width: 100%;
-  cursor: pointer;
-
-  img {
-    transition: all .2s ease-in;
-    &:hover {
-      transform: scale(1.01);
-      box-shadow: var(--luxe-box-shadow);
-    }
-  }
-}
-
-.galleryItems {
-  padding: 16px 0;
-  margin: 0;
-  display: flex;
-  flex-direction: row;
-  gap: 16px;
-  list-style: none;
-
-  li {
-    display: flex;
-  }
-}
-
-.galleryMore,
-.galleryItem {
-  flex: 1;
-  min-width: 0;
-  min-height: 0;
-}
-
-.galleryMore,
-.galleryItem {
-  /* TODO: Same style than card  */
-  border-radius: 16px;
-  transition: all .2s ease-in;
-  margin-bottom: 0;
-  cursor: pointer;
-
-  &:hover {
-    transform: scale(1.05);
-    box-shadow: var(--luxe-box-shadow);
-  }
-}
-
-.galleryMore {
-  cursor: pointer;
-  display: flex;
-  background-color: var(--luxe-color-gray);
-  align-items: center;
-  justify-content: center;
-  font-size: 23px;
-  font-weight: bold;
-  border-radius: 16px;
-}
-
-
-/* Gallery */
-.dialogContentImage {
-  display: flex;
-  gap: 16px;
-  margin: 20px;
-}
-
-.dialogButton {
-  border: 0;
-  background: transparent;
-  color: var(--luxe-color-gray);
-  opacity: 0.6;
-  transition: color 0.2s ease-in-out;
-  cursor: pointer;
-  font-size: 40px;
-
-  &:focus {
-    outline: none;
-  }
-
-  &:focus-visible {
-    outline: 2px solid var(--luxe-color-gold);
-    outline-offset: 2px;
-  }
-
-  &:hover {
-    opacity: 1;
-    color: var(--luxe-color-gold);
-  }
-
-  &[disabled] {
-    opacity: 0.2;
-    pointer-events: none;
-  }
-}
-
-.dialogButtonNav {
-  width: 80px;
-  font-size: 120px;
-}
-
-.dialogButtonClose {
-  position: fixed;
-  top: 24px;
-  right: 24px;
-  width: 40px;
-  height: 40px;
-  padding: 0;
-  text-align: center;
-  line-height: 1;
-}
-
-.dialogInfo {
-  color: var(--luxe-color-gray);
-  font-size: 13px;
-  text-align: right;
-  margin-right: 108px;
-}
-
-
-.dialogImage {
-  height: 80vh;
-}
