--- src/commons/Picture.module.css
@@ -1,3 +1,0 @@
-.image {
-  border-radius: 0;
-}
