+++ src/commons/Picture.tsx
@@ -1,25 +1,12 @@
-import clsx from "clsx";
-import classes from "./Picture.module.css";
+import type { PictureProps } from "~/commons/types";
 
-interface PictureProps {
-  image: { src: string; alt: string };
-  sources?: { media: string; srcSet: string }[];
-  height?: string;
-  className?: string;
-}
-
-export const Picture = ({ image, sources, height, className }: PictureProps) => {
-  return (
-    <picture>
-      {sources?.map((source) => (
-        <source key={source.media} media={source.media} srcSet={source.srcSet} />
-      ))}
-      <img
-        src={image.src}
-        alt={image.alt}
-        className={clsx(classes.image, className)}
-        height={height}
-      />
-    </picture>
-  );
+export const Picture = ({ image, sources, height, className, onClick }: PictureProps) => {
+	return (
+		<picture onClick={onClick} className={className}>
+			{sources?.map((source) => (
+				<source key={source.media} media={source.media} srcSet={source.srcSet} />
+			))}
+			<img src={image.src} alt={image.alt} height={height} />
+		</picture>
+	);
 };
