+++ src/commons/gallery/Gallery.client.module.css
@@ -1,0 +1,113 @@
+@value variables: "~/templates/css/variables.css";
+
+@value gridBreakpoints_sm, gridBreakpoints_md, gridBreakpoints_xl, space_3, space_4 from variables;
+
+/*@keyframes fadeInListItem {*/
+/*	from { opacity: 0; transform: translateY(20px);}*/
+/*	to   { opacity: 1; transform: translateY(0);}*/
+/*}*/
+/*@keyframes fadeOutListItem {*/
+/*	from { opacity: 1; transform: translateY(0);}*/
+/*	to   { opacity: 0; transform: translateY(-20px);}*/
+/*}*/
+
+/*!* Par défaut, animé à l’entrée *!*/
+/*.item {*/
+/*	animation: fadeInListItem 0.3s;*/
+/*}*/
+
+/*!* Classe à ajouter lors de la suppression *!*/
+/*.itemLeaving {*/
+/*	animation: fadeOutListItem 0.3s forwards;*/
+/*}*/
+
+.dialog {
+	--luxe-dialog-opacity: 1;
+}
+
+.picture {
+  margin-bottom: 0;
+  width: 100%;
+  cursor: pointer;
+
+  img {
+    transition: all .2s ease-in;
+    &:hover {
+      transform: scale(1.01);
+      box-shadow: var(--luxe-box-shadow);
+    }
+  }
+}
+
+.thumbnails {
+	display: flex;
+	flex-direction: row;
+  padding: 16px 0;
+  margin: 0;
+  gap: space_3;
+  list-style: none;
+
+  li {
+    display: flex;
+		flex: 0 1 100%;
+		max-width: 100%;
+		@media (min-width: gridBreakpoints_sm) {
+			& {
+				flex: 0 1 50%;
+				max-width: 50%;
+			}
+		}
+		@media (min-width: gridBreakpoints_md) {
+			& {
+				flex: 0 1 33.3333%;
+				max-width: 33.3333%;
+			}
+		}
+		/*@media (min-width: gridBreakpoints_lg) {*/
+		/*	& {*/
+		/*		flex: 0 1 50%;*/
+		/*		max-width: 50%;*/
+		/*	}*/
+		/*}*/
+		@media (min-width: gridBreakpoints_xl) {
+			& {
+				flex: 0 1 20%;
+				max-width: 20%;
+			}
+		}
+  }
+}
+
+.more,
+.item {
+  flex: 1;
+  min-width: 0;
+  min-height: 0;
+}
+
+.more,
+.item {
+	composes: card from "~/templates/css/global.module.css";
+	padding: 0 !important;
+  /*!* TODO: Same style than card  *!*/
+  /*border-radius: 16px;*/
+  /*transition: all .2s ease-in;*/
+  /*margin-bottom: 0;*/
+  /*cursor: pointer;*/
+
+  /*&:hover {*/
+  /*  transform: scale(1.05);*/
+  /*  box-shadow: var(--luxe-box-shadow);*/
+  /*}*/
+}
+
+.more {
+  cursor: pointer;
+  display: flex;
+  background-color: var(--luxe-color-gray);
+  align-items: center;
+  justify-content: center;
+  font-size: 23px;
+  font-weight: bold;
+  border-radius: 16px;
+}
