+++ src/commons/gallery/Slideshow.client.module.css
@@ -1,0 +1,118 @@
+@value variables: "~/templates/css/variables.css";
+
+@value gridBreakpoints_sm, space_3, space_4 from variables;
+
+
+.contentImage {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: center;
+	gap: space_3;
+	margin: space_3;
+
+	picture{
+		flex: 1;
+		text-align: center;
+	}
+
+	img {
+		max-width: 90vw;
+		max-height: 50vh;
+		width: 100%;
+		height: 100%;
+	}
+
+	@media (min-width: gridBreakpoints_sm) {
+		& {
+			flex-direction: row;
+			gap: space_4;
+			margin: space_4;
+			img {
+				/*max-width: 70vw;*/
+				max-height: 80vh;
+			}
+		}
+	}
+}
+
+.button {
+	border: 0;
+	background: transparent;
+	color: var(--luxe-color-gray);
+	opacity: 0.6;
+	transition: color 0.2s ease-in-out;
+	cursor: pointer;
+	font-size: 32px;
+	line-height: 1;
+
+	&:focus {
+		outline: none;
+	}
+	&:focus-visible {
+		outline: 2px solid var(--luxe-color-gold);
+		outline-offset: 2px;
+	}
+	&:hover {
+		opacity: 1;
+		color: var(--luxe-color-gold);
+	}
+	&[disabled] {
+		opacity: 0.2;
+		pointer-events: none;
+	}
+}
+
+
+.buttonNav {
+	width: 44px;
+	font-size: 60px;
+
+	@media (min-width: gridBreakpoints_sm) {
+		& {
+			width: 80px;
+			font-size: 120px;
+		}
+	}
+}
+
+.buttonClose {
+	position: fixed;
+	top: 10px;
+	right: 10px;
+	width: 36px;
+	height: 36px;
+	font-size: 28px;
+	padding: 0;
+	text-align: center;
+	line-height: 1;
+	z-index: 1001;
+	@media (min-width: gridBreakpoints_sm) {
+		& {
+			top: 24px;
+			right: 24px;
+			width: 40px;
+			height: 40px;
+			font-size: 40px;
+		}
+	}
+}
+
+.info {
+	color: var(--luxe-color-gray);
+	font-size: 13px;
+	text-align: center;
+	margin: space_3 0 0 0;
+	word-break: break-word;
+	padding: 0 8vw;
+
+	@media (min-width: gridBreakpoints_sm) {
+		& {
+			text-align: right;
+			margin: 0 108px 0 0;
+			padding: 0;
+			font-size: 15px;
+		}
+	}
+}
+
