+++ src/commons/index.ts
@@ -1,10 +1,9 @@
 export * from "./PageTitle";
 export * from "./Figure";
-// export * from "./Gallery.client";
 export * from "./ContentHeader";
 export * from "./List";
 export * from "./HeadingSection";
 export * from "./grid";
-export * from "./CMPreview";
 export * from "./Contact";
 export * from "./Picture";
+export * from "./types";
