+++ src/components/Realtor/cm.server.tsx
@@ -1,16 +1,16 @@
 import { jahiaComponent, Render } from "@jahia/javascript-modules-library";
-import { CMPreview } from "~/commons";
+import { CMPreview } from "~/templates/CMPreview";
 
 jahiaComponent(
-  {
-    nodeType: "luxe:realtor",
-    name: "cm",
-    displayName: "jContent internal view",
-    componentType: "view",
-  },
-  (_, { currentNode }) => (
-    <CMPreview>
-      <Render node={currentNode} view="fullPage" />
-    </CMPreview>
-  ),
+	{
+		nodeType: "luxe:realtor",
+		name: "cm",
+		displayName: "jContent internal view",
+		componentType: "view",
+	},
+	(_, { currentNode }) => (
+		<CMPreview>
+			<Render node={currentNode} view="fullPage" />
+		</CMPreview>
+	),
 );
