+++ src/templates/CMPreview.tsx
@@ -1,0 +1,9 @@
+import { AddResources, buildModuleFileUrl } from "@jahia/javascript-modules-library";
+import type { CommonsProps } from "~/commons/types";
+
+export const CMPreview = ({ className, children }: CommonsProps) => (
+	<>
+		<AddResources type="css" resources={buildModuleFileUrl("dist/assets/style.css")} />
+		<main className={className}>{children}</main>
+	</>
+);
